*** empty log message ***

+++ install-scripts/linux/copyLogsToWeb.sh
@@ -6,19 +6,24 @@ W=/home/orange/estelleDownload/regressionResults
 rm $W/linux.*.log
 
 # copy new ones
-cd /home/orange/daily/orange
-cp compiling.log $W/linux.compiling.log
-cp ../output.log $W/linux.output.log
-cp ../regress.log $W/linux.regress.log
+if cd /home/orange/daily/orange ; then
+	cp compiling.log $W/linux.compiling.log
+	cp ../output.log $W/linux.output.log
+	cp ../regress.log $W/linux.regress.log
+else
+	echo "no log files"
+fi
 
-cd /home/orange/daily/test_install/orange
-
-for f in *-output; do
-	echo copying $f
-	mkdir $W/$f
-	rm $W/$f/*linux2.*.txt
-	cp $f/*crash.txt $f/*error.txt $f/*new.txt $f/*changed.txt $f/*random1.txt $f/*random2.txt $W/$f
-done
+if cd /home/orange/daily/test_install/orange ; then
+	for f in *-output; do
+		echo copying $f
+		mkdir $W/$f
+		rm $W/$f/*linux2.*.txt
+		cp $f/*crash.txt $f/*error.txt $f/*new.txt $f/*changed.txt $f/*random1.txt $f/*random2.txt $W/$f
+	done
+else
+	echo "regression test were not run"
+fi
 
 /home/orange/umount_estelleDownload
